refactor: cambiar la ruta protegida a la carpeta de organismos
