+++ src/App.jsx
@@ -21,7 +21,7 @@ import SetsDeProductos from "./vistas/paginas/setsProducto";
 import Eventos from "./vistas/paginas/eventos";
 import Configuracion from "./vistas/paginas/configuracion";
 import { AuthProvider } from "./hooks/AuthProvider";
-import RutaProtegida from "./RutaProtegida";
+import RutaProtegida from "./vistas/componentes/organismos/RutaProtegida";
 import PaginaInicioSesion from "./vistas/paginas/inicioSesion";
 import Tienda from "./Tienda";
